Update the  type from ll l to make -Werror happy

+++ 0x04-more_functions_nested_loops/main.h
@@ -8,6 +8,6 @@ int mul(int a, int b);
 void print_numbers(void);
 void print_most_numbers(void);
 void more_numbers(void);
-void print_chars(unsigned long long int);
+void print_chars(unsigned long int);
 
 #endif
